add mTLS, reuse getConfig func
