+++ test/e2e_env/universal/envoyconfig/builtingateway.go
@@ -1,12 +1,10 @@
 package envoyconfig
 
 import (
-	"encoding/json"
 	"fmt"
 	"net"
 	"os"
 	"path/filepath"
-	"slices"
 	"strconv"
 	"strings"
 
@@ -14,13 +12,10 @@ import (
 	. "github.com/onsi/gomega"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	"github.com/kumahq/kuma/api/openapi/types"
-	api_common "github.com/kumahq/kuma/api/openapi/types/common"
 	meshtimeout "github.com/kumahq/kuma/pkg/plugins/policies/meshtimeout/api/v1alpha1"
 	"github.com/kumahq/kuma/pkg/test"
 	"github.com/kumahq/kuma/pkg/test/matchers"
 	"github.com/kumahq/kuma/pkg/test/resources/builders"
-	"github.com/kumahq/kuma/pkg/util/pointer"
 	. "github.com/kumahq/kuma/test/e2e_env/universal/gateway"
 	. "github.com/kumahq/kuma/test/framework"
 	"github.com/kumahq/kuma/test/framework/envs/universal"
@@ -84,9 +79,11 @@ spec:
 					builders.Mesh().
 						WithName(mesh).
 						WithoutInitialPolicies().
-						WithMeshServicesEnabled(mesh_proto.Mesh_MeshServices_Exclusive),
+						WithMeshServicesEnabled(mesh_proto.Mesh_MeshServices_Exclusive).
+						WithBuiltinMTLSBackend("ca-1").WithEnabledMTLSBackend("ca-1"),
 				),
 			).
+			Install(MeshTrafficPermissionAllowAllUniversal(mesh)).
 			Install(GatewayClientAppUniversal("gateway-client")).
 			Install(EchoServerApp(mesh, "echo-server", "echo-service", "universal")).
 			Install(GatewayProxyUniversal(mesh, "gateway-proxy")).
@@ -94,6 +91,8 @@ spec:
 			Install(YamlUniversal(meshHTTPRoute()))
 
 		Expect(setup.Setup(universal.Cluster)).To(Succeed())
+
+		waitMeshServiceReady(mesh, "echo-service")
 
 		Eventually(ProxySimpleRequests(universal.Cluster, "universal",
 			gatewayAddressPort("gateway-proxy", gatewayPort), "example.kuma.io"), "60s", "1s").Should(Succeed())
@@ -117,27 +116,6 @@ spec:
 		)).To(Succeed())
 	})
 
-	getConfig := func(dpp string) string {
-		output, err := universal.Cluster.GetKumactlOptions().
-			RunKumactlAndGetOutput("inspect", "dataplane", dpp, "--type", "config", "--mesh", mesh, "--shadow", "--include=diff")
-		Expect(err).ToNot(HaveOccurred())
-		redacted := redactStatPrefixes(redactIPs(output))
-
-		response := types.InspectDataplanesConfig{}
-		Expect(json.Unmarshal([]byte(redacted), &response)).To(Succeed())
-		Expect(response.Diff).ToNot(BeNil())
-		response.Diff = pointer.To(slices.DeleteFunc(*response.Diff, func(item api_common.JsonPatchItem) bool {
-			return item.Op == api_common.Test
-		}))
-		slices.SortStableFunc(*response.Diff, func(a, b api_common.JsonPatchItem) int {
-			return strings.Compare(a.Path, b.Path)
-		})
-
-		result, err := json.MarshalIndent(response, "", "  ")
-		Expect(err).ToNot(HaveOccurred())
-		return string(result)
-	}
-
 	DescribeTable("should generate proper Envoy config",
 		func(inputFile string) {
 			// given
@@ -150,7 +128,7 @@ spec:
 			}
 
 			// then
-			Expect(getConfig("gateway-proxy")).To(matchers.MatchGoldenJSON(strings.Replace(inputFile, "input.yaml", "gateway-proxy.golden.json", 1)))
+			Expect(getConfig(mesh, "gateway-proxy")).To(matchers.MatchGoldenJSON(strings.Replace(inputFile, "input.yaml", "gateway-proxy.golden.json", 1)))
 		},
 		test.EntriesForFolder(filepath.Join("builtingateway", "meshtimeout"), "envoyconfig"),
 	)
